Use `gatsby-transformer-csv` to provide data
