+++ .gitignore
@@ -1,2 +1,3 @@
 .cache
 node_modules
+public
