+++ gatsby-config.js
@@ -21,5 +21,18 @@ module.exports = {
       },
       __key: "images",
     },
+    {
+      resolve: "gatsby-source-filesystem",
+      options: {
+        name: "data",
+        path: `${__dirname}/src/data/`,
+      },
+    },
+    {
+      resolve: `gatsby-transformer-csv`,
+      options: {
+        headers: ["question", "important", "response"],
+      },
+    },
   ],
 };
