--- public/page-data/404.html/page-data.json
@@ -1,1 +1,0 @@
-{"componentChunkName":"component---src-pages-404-js","path":"/404.html","result":{"pageContext":{}},"staticQueryHashes":[]}
