--- public/page-data/cards/page-data.json
@@ -1,1 +1,0 @@
-{"componentChunkName":"component---src-pages-cards-js","path":"/cards/","result":{"pageContext":{}},"staticQueryHashes":[]}
