--- public/page-data/dev-404-page/page-data.json
@@ -1,1 +1,0 @@
-{"componentChunkName":"component---cache-dev-404-page-js","path":"/dev-404-page/","result":{"data":{"allSitePage":{"nodes":[{"path":"/404/"},{"path":"/cards/"},{"path":"/"},{"path":"/404.html"}]}},"pageContext":{}},"staticQueryHashes":[]}
