--- public/page-data/index/page-data.json
@@ -1,1 +1,0 @@
-{"componentChunkName":"component---src-pages-index-js","path":"/","result":{"pageContext":{}},"staticQueryHashes":[]}
