+++ src/data/data.csv
@@ -1,3 +1,4 @@
+PHRASE,IMPORTANT++,BONNE REPONSE
 La qualité sans compromis !,OUI,❤
 Data Scientist love Python,OUI,❤
 Les polos violets,NON,❤
