+++ src/data/data.js
@@ -1,182 +1,21 @@
-export default [
-  {
-    question: "La qualité sans compromis !",
-    important: true,
-    response: true,
-  },
-  {
-    question: "Data Scientist love Python",
-    important: true,
-    response: true,
-  },
-  {
-    question: "Les polos violets",
-    important: false,
-    response: true,
-  },
-  {
-    question: "Les tisanes au houblon",
-    important: false,
-    response: true,
-  },
-  {
-    question: '"Etre agile',
-    important: false,
-    response: false,
-  },
-  {
-    question: "Cycle en V",
-    important: true,
-    response: false,
-  },
-  {
-    question: "Le pair programming c'est une perte de temps",
-    important: true,
-    response: false,
-  },
-  {
-    question: "Tester c'est douter !",
-    important: true,
-    response: false,
-  },
-  {
-    question: "Mobile first",
-    important: false,
-    response: true,
-  },
-  {
-    question: "Storybook ne sert à rien",
-    important: true,
-    response: false,
-  },
-  {
-    question: "Design system",
-    important: true,
-    response: true,
-  },
-  {
-    question: "TypeScript pour le typage",
-    important: true,
-    response: true,
-  },
-  {
-    question: "Vive le Serverless",
-    important: true,
-    response: true,
-  },
-  {
-    question: "Oui à l'UI as Code",
-    important: true,
-    response: true,
-  },
-  {
-    question: "SASS over CSS",
-    important: true,
-    response: true,
-  },
-  {
-    question: "Automatiser les tests fonctionnels",
-    important: true,
-    response: true,
-  },
-  {
-    question: '"Des fois',
-    important: false,
-    response: false,
-  },
-  {
-    question: "La programmation réactive !",
-    important: true,
-    response: true,
-  },
-  {
-    question: "La programmation fonctionnelle !",
-    important: true,
-    response: true,
-  },
-  {
-    question: "NetBeans l'IDE du futur",
-    important: false,
-    response: false,
-  },
-  {
-    question: "Ecologie !",
-    important: false,
-    response: true,
-  },
-  {
-    question: "Gérôme Egron",
-    important: true,
-    response: true,
-  },
-  {
-    question: "En route pour Devoxx",
-    important: true,
-    response: true,
-  },
-  {
-    question: "DRY",
-    important: true,
-    response: true,
-  },
-  {
-    question: "Une estimation est avant tout précise.",
-    important: true,
-    response: false,
-  },
-  {
-    question: "Expert après un Hello World",
-    important: false,
-    response: false,
-  },
-  {
-    question: "J'aime que mon code soit SOLID",
-    important: true,
-    response: true,
-  },
-  {
-    question: 'Espaces ou tabulations',
-    important: false,
-    response: false,
-  },
-  {
-    question: '"100% de couverture de code',
-    important: false,
-    response: false,
-  },
-  {
-    question: "L'équipe ? Oui je crois qu'on est plusieurs sur le projet.",
-    important: true,
-    response: false,
-  },
-  {
-    question: "T'as fait un ticket ?",
-    important: false,
-    response: false,
-  },
-  {
-    question: 'Le nouveau galère',
-    important: false,
-    response: false,
-  },
-  {
-    question: "J'allume mes serveurs seulement quand c'est nécessaire.",
-    important: true,
-    response: true,
-  },
-  {
-    question: "Tu veux qu'on regarde ensemble ?",
-    important: true,
-    response: true,
-  },
-  {
-    question: "La revue de code est un défouloir.",
-    important: true,
-    response: false,
-  },
-  {
-    question: "Partage de la connaissance",
-    important: true,
-    response: true,
-  },
-];
+import * as React from "react";
+import { graphql, StaticQuery } from "gatsby";
+
+import buildDataset from './utils/buildDataset';
+
+export default (Component) => () => (
+  <StaticQuery
+    query={graphql`
+      query AllData {
+        allDataCsv {
+          nodes {
+            question
+            important
+            response
+          }
+        }
+      }
+    `}
+    render={(allData) => <Component data={buildDataset(allData.allDataCsv.nodes)}></Component>}
+  />
+);
